Separate references file
Separate references into another file. The anchor usage wasn't very
readable as text and the linking didn't scroll far enough down because
it didn't take into account the GitHub website header space. So, let's
put the references in another file. It also means that the README is a
bit shorter which is desirable to a degree.